M253 remove hack
